LayoutTests/media/media-extension-with-fragment-expected.txt

EVENT(loadeddata)
MediaPlayer extracted extension from src url. OK

END OF TEST
